refactor(objectives): move Evidence materialization into analysis
Why:
- Durable Evidence construction and paper accounting remained hidden in the coordinator after source facts had already moved to direct research-stage owners.

What:
- Move draft selection, Source resolution, Source-keyed Evidence construction, duplicate resolution, and PaperContribution accounting to one direct analysis owner.

Impact:
- Evidence identities, provenance, ordering, scientific states, paper dispositions, persistence inputs, and public contracts remain unchanged without a forwarding path.

Verification:
- All 131 targeted Objective analysis and publication tests passed, and compilation, scoped lint, documentation governance, diff, formatting, and ownership checks completed successfully.
